requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0
django-axes==6.3.0
django-defender==0.9.7
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
Pillow==10.1.0
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
dj-database-url==0.5.0
psycopg2==2.9.7
gunicorn==20.1.0
django-environ>=0.8.1
whitenoise==6.2.0
boto3==1.26.15
django-storages==1.13.2
